+++ wo/cli/plugins/clean.py
@@ -74,18 +74,18 @@ class WOCleanController(CementBaseController):
     def clean_opcache(self):
         if (os.path.exists('/usr/sbin/nginx') and
                 os.path.exists(
-                    '/var/www/22222/htdocs/cache/opcache/opgui.php')):
+                    '/var/www/22222/htdocs/cache/opcache/php72.php')):
             try:
                 Log.info(self, "Cleaning opcache")
                 opgui = requests.get(
-                    "https://127.0.0.1:22222/cache/opcache/opgui.php?reset=1")
-                if opgui.status_code != '200':
+                    "http://127.0.0.1/cache/opcache/php72.php")
+                if opgui.status_code != '200' or opgui.status_code != '302':
                     Log.warn(self, 'Cleaning opcache failed')
             except Exception as e:
                 Log.debug(self, "{0}".format(e))
                 Log.debug(self, "Unable hit url, "
-                          " https://127.0.0.1:22222/cache/opcache/"
-                          "opgui.php?reset=1,"
+                          " http://127.0.0.1/cache/opcache/"
+                          "php72.php,"
                           " please check you have admin tools installed")
                 Log.debug(self, "please check you have admin tools installed,"
                           " or install them with `wo stack install --admin`")